Memory usage optimization: change from IntVector to vector<short int> to store states of sequences

+++ alignment/alisimulatorinvar.cpp
@@ -57,7 +57,7 @@ void AliSimulatorInvar::simulateSeqs(int sequence_length, double *site_specific_
         node->num_children_done_simulation++;
         // remove the sequence of
         if (!node->isLeaf() && node->num_children_done_simulation >= (node->neighbors.size() - 1))
-            vector<int>().swap(node->sequence);
+            vector<short int>().swap(node->sequence);
         
         // browse 1-step deeper to the neighbor node
         simulateSeqs(sequence_length, site_specific_rates, model, trans_matrix, max_num_states, (*it)->node, node);
